Refactored data class to appropriate folder
