+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/popup/ContinentDialog.kt
@@ -6,22 +6,23 @@ import android.os.Bundle
 import android.widget.*
 import androidx.fragment.app.DialogFragment
 import com.se2gruppe5.risikofrontend.R
+import com.se2gruppe5.risikofrontend.game.dataclasses.ContinentInfoRecord
 
 
 @SuppressWarnings("unused")
 class ContinentDialog : DialogFragment(){
 
     val continentList = listOf(
-        ContinentInfo("Power Supply", "#b5edf9", 6, 2),
-        ContinentInfo("MMC", "#e8f3a8", 2, 2),
-        ContinentInfo("RAM", "#87deb3", 10, 2),
-        ContinentInfo("DCON", "#f0a8e1", 5, 2),
-        ContinentInfo("CPU", "#e6afaf", 6, 2),
-        ContinentInfo("Essentials", "#fec466", 7, 2),
-        ContinentInfo("Southbridge", "#f5d557", 4, 2),
-        ContinentInfo("Wireless Mesh", "#9effad", 8, 2),
-        ContinentInfo("Embedded Controller", "#9b3e3e", 7, 2),
-        ContinentInfo("CMOS", "#a4e5e2", 5, 2)
+        ContinentInfoRecord("Power Supply", "#b5edf9", 6, 2),
+        ContinentInfoRecord("MMC", "#e8f3a8", 2, 2),
+        ContinentInfoRecord("RAM", "#87deb3", 10, 2),
+        ContinentInfoRecord("DCON", "#f0a8e1", 5, 2),
+        ContinentInfoRecord("CPU", "#e6afaf", 6, 2),
+        ContinentInfoRecord("Essentials", "#fec466", 7, 2),
+        ContinentInfoRecord("Southbridge", "#f5d557", 4, 2),
+        ContinentInfoRecord("Wireless Mesh", "#9effad", 8, 2),
+        ContinentInfoRecord("Embedded Controller", "#9b3e3e", 7, 2),
+        ContinentInfoRecord("CMOS", "#a4e5e2", 5, 2)
     )
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         val builder = AlertDialog.Builder(requireContext())
